IN-905 Rename db.py to database.py
